+++ extra_tests/snippets/builtin_hash.py
@@ -1,3 +1,5 @@
+import sys
+
 from testutils import assert_raises
 
 
@@ -28,3 +30,19 @@ with assert_raises(TypeError):
 
 with assert_raises(TypeError):
     hash([])
+
+# Hashing a deeply nested tuple must not run off the native stack: the hash
+# slot dispatch is what recurses, so that is where the depth is checked.
+
+if sys.implementation.name == "rustpython":
+    # CPython, which also runs this snippet, survives this depth unguarded.
+    deep_tuple = ()
+    for _ in range(sys.getrecursionlimit() * 2):
+        deep_tuple = (deep_tuple,)
+    with assert_raises(RecursionError):
+        hash(deep_tuple)
+    # a dict key and a set member are hashed on insertion, same dispatch
+    with assert_raises(RecursionError):
+        {deep_tuple: 1}
+    with assert_raises(RecursionError):
+        {deep_tuple}
