+++ extra_tests/snippets/stdlib_imp.py
@@ -1,6 +1,8 @@
 import _imp
 import time as import_time
 
+from testutils import assert_raises
+
 assert _imp.is_builtin("time") == True
 assert _imp.is_builtin("os") == False
 assert _imp.is_builtin("not existing module") == False
@@ -29,3 +31,8 @@ _imp.get_frozen_object("__hello__")
 
 hello = _imp.init_frozen("__hello__")
 assert hello.initialized == True
+
+# withdata is keyword-only
+with assert_raises(TypeError):
+    _imp.find_frozen("x", True)
+assert _imp.find_frozen("_this_module_does_not_exist_") is None
