+++ extra_tests/snippets/stdlib_pwd.py
@@ -12,3 +12,7 @@ with assert_raises(KeyError):
     fake_name = "fake_user"
     while pwd.getpwnam(fake_name):
         fake_name += "1"
+
+# The field getters must not index a struct sequence that __new__ never filled.
+with assert_raises(TypeError):
+    pwd.struct_passwd()
