+++ extra_tests/snippets/stdlib_traceback.py
@@ -1,4 +1,8 @@
+import itertools
 import traceback
+
+import _suggestions
+from testutils import assert_raises
 
 try:
     1 / 0
@@ -25,3 +29,10 @@ try:
 except ZeroDivisionError as ex2:
     tb = traceback.extract_tb(ex2.__traceback__)
     assert len(tb) == 1
+
+# The candidate list backing "Did you mean" suggestions is a list; an arbitrary
+# iterable must be rejected rather than drained.
+
+with assert_raises(TypeError):
+    _suggestions._generate_suggestions(itertools.count(), "x")
+assert _suggestions._generate_suggestions(["value"], "valu") == "value"
