+++ extra_tests/snippets/stdlib_types.py
@@ -1,5 +1,6 @@
 import _ast
 import platform
+import sys
 import types
 
 from testutils import assert_raises
@@ -34,3 +35,26 @@ def _run_missing_type_params_regression():
 
 
 _run_missing_type_params_regression()
+
+if sys.implementation.name == "rustpython":
+    # __parameters__ is computed when the alias is built, and the walk descends
+    # into every list and tuple argument, so a self-referential or deeply
+    # nested argument must be caught. CPython, which also runs this snippet,
+    # does not walk into plain lists at all.
+    self_referential = []
+    self_referential.append(self_referential)
+    with assert_raises(RecursionError):
+        list[self_referential]
+
+    nested = [0]
+    for _ in range(sys.getrecursionlimit() * 2):
+        nested = [nested]
+    with assert_raises(RecursionError):
+        list[nested]
+
+    # hashing an alias walks the same shape
+    deep_alias = int
+    for _ in range(sys.getrecursionlimit() * 2):
+        deep_alias = list[deep_alias]
+    with assert_raises(RecursionError):
+        hash(deep_alias)
